Extract proposal data to separate file
